Renamed `Parser` to `NewspeakParser`
